requirements.txt
streamlit==1.47.0
pdfplumber==0.11.7
pycountry==24.6.1
google-generativeai==0.8.5
